new order struct and items (#15)
* itemNumber is null

https://trello.com/c/MXsUgx6B/19-%D0%BF%D0%BE%D0%BB%D0%B5-note-%D0%B4%D0%BB%D1%8F-orders-%D0%B8-items-%D0%B4%D0%BE%D0%BB%D0%B6%D0%BD%D0%BE-%D0%B1%D1%8B%D1%82%D1%8C-%D1%80%D0%B5%D0%B4%D0%B0%D0%BA%D1%82%D0%B8%D1%80%D1%83%D0%B5%D0%BC%D1%8B%D0%BC-%D0%B8%D0%B7-%D1%82%D0%B0%D0%B1%D0%BB%D0%B8%D1%86%D1%8B

* change filter name

* ad ai service and state request

* fix

* ai api added

* new order struct and items

* fix dto

* читабельные ошибки при логине

* рефакторинг, добавил типы, выпилил неиспользуемый код

* рефач, удалил лишнее

* fix

* добавил редьюсеры

* проверка что все items в заказе валидные

* фиксы

* подгрузка значений по SKU

* delete migr

* fix error message

* убрал лишнее

* сделал подгрузку статусов по заказов по интервалу

* i > c

* @IsDate()

* fix джобы в гитхабе

* добавил healthcheck

* fix

+++ src/grainger-accounts/grainger-accounts.service.ts
@@ -16,11 +16,11 @@ export class GraingerAccountsService {
     return this.repository.find();
   }
 
-  getById(id: string): Promise<GraingerAccount> {
-    return this.repository.findOne(id);
+  getById(id: string, where?: any): Promise<GraingerAccount> {
+    return this.repository.findOne(id, { ...where });
   }
 
-  async add(accountDto: CreateGraingerAccountDto) {
+  async add(accountDto: CreateGraingerAccountDto): Promise<GraingerAccount> {
     const accountWithLogin = await this.repository.findOne({
       where: { email: accountDto.email },
     });
@@ -31,14 +31,10 @@ export class GraingerAccountsService {
       );
     }
 
-    let account = accountDto;
-    if (!(accountDto instanceof GraingerAccount)) {
-      account = GraingerAccount.create(accountDto);
-    }
-    return await this.repository.save(account);
+    return this.repository.save(GraingerAccount.create(accountDto));
   }
 
-  async editById(id: string, dto: any) {
+  async editById(id: string, dto: any): Promise<GraingerAccount> {
     const accountWithLogin = await this.repository.findOne({
       where: { email: dto.login },
     });
@@ -55,13 +51,16 @@ export class GraingerAccountsService {
     }
 
     const updated = Object.assign(account, dto);
-    return await this.repository.save(updated);
+    return this.repository.save(updated);
   }
 
   async deleteById(id: string) {
-    const account = await this.getById(id);
+    const account = await this.getById(id, { relations: ['items'] });
     if (!account) {
       throw new HttpException(`Grainger Account doesn't exist`, HttpStatus.OK);
+    }
+    if (account.items?.length) {
+      throw new HttpException(`Grainger Account already have items`, HttpStatus.OK);
     }
     return this.repository.delete(id);
   }
